+++ CHANGELOG.md
@@ -1,5 +1,56 @@
 # History of changes
 
+## Version 0.4.7 (2018-01-08)
+
+### Issues Closed
+
+#### New Features
+
+* [Issue 147](https://github.com/OpenWaterAnalytics/pyswmm/issues/147) - Add before_end simulation hook.
+
+In this release 1 issue was closed.
+
+### Pull Requests Merged
+
+* [PR 148](https://github.com/OpenWaterAnalytics/pyswmm/pull/148) - Added additional callback hook for Before_End
+
+In this release 1 pull request was closed.
+
+## Version 0.4.6 (2018-01-02)
+
+### Issues Closed
+
+#### New Features
+
+* [Issue 140](https://github.com/OpenWaterAnalytics/pyswmm/issues/140) - Add simulation function hook for callbacks
+
+#### Bugs fixed
+
+* [Issue 143](https://github.com/OpenWaterAnalytics/pyswmm/issues/143) - Remove bare exceptions
+
+In this release 2 issues were closed.
+
+### Pull Requests Merged
+
+* [PR 144](https://github.com/OpenWaterAnalytics/pyswmm/pull/144) - Added swmm_lib_path argument to Simulation and PySWMM class
+* [PR 142](https://github.com/OpenWaterAnalytics/pyswmm/pull/142) - Added Hooks for Callbacks
+
+In this release 2 pull requests were closed.
+
+## Version 0.4.5 (2017-12-27)
+
+### Issues Closed
+
+#### New Features
+
+* [Issue 90](https://github.com/OpenWaterAnalytics/pyswmm/issues/90) - Add test *.inp files to distribution
+
+#### Bugs fixed
+
+* [Issue 139](https://github.com/OpenWaterAnalytics/pyswmm/issues/139) - Update SWMM5 Lib Windows - Linking Issue
+* [Issue 136](https://github.com/OpenWaterAnalytics/pyswmm/issues/136) - broken doctests
+
+In this release 3 issues were closed.
 
 ## Version 0.4.4 (2017-11-14)